+++ @planetarium/account/src/Account.ts
@@ -4,13 +4,13 @@ import RawPrivateKey from "./RawPrivateKey.js";
 import Signature from "./Signature.js";
 
 export interface Account {
-  publicKey: PublicKey;
+  getPublicKey(): Promise<PublicKey>;
   sign(message: Message): Promise<Signature>;
 }
 
 export function isAccount(account: unknown): account is Account {
   return (
-    (account as { publicKey: unknown }).publicKey instanceof PublicKey &&
+    (account as { getPublicKey: unknown }).getPublicKey instanceof Function &&
     (account as { sign: unknown }).sign instanceof Function
   );
 }