+++ CHANGES.md
@@ -6,6 +6,13 @@ Version 0.54.0
 
 To be released.
 
+### Deprecated APIs
+
+ -  (@planerarium/account)  Deprecated `RawPrivateKey.publicKey` property in
+    favour of `RawPrivateKey.getPublicKey()` async method.  [[#3061]]
+ -  (@planetarium/account-aws-kms)  Deprecated `AwsKmsAccount.publicKey` in
+    favour of `AwsKmsAccount.getPublicKey()` async method.  [[#3061]]
+
 ### Backward-incompatible API changes
 
  -  Removed `TxMetadata` class.  [[#1164], [#2986]]
@@ -58,6 +65,11 @@ To be released.
  -  Removed `PreEvaluationBlock<T>.Mine()` and `BlockMetadata.MineNonce()`
     methods.  [[#3067]]
  -  Removed `HashCash` class.  [[#3067]]
+ -  (@planetarium/account) Replaced `Account.publicKey` property with
+    `Account.getPublicKey()` async method.  [[#3061]]
+ -  (@planetarium/account) `Address.deriveFrom()` method now returns
+    `Promise<Address>` when an `Account` is given.  However, it still returns
+    `Address` when a `PublicKey` is given.  [[#3061]]
 
 ### Backward-incompatible network protocol changes
 
@@ -82,6 +94,12 @@ To be released.
  -  Added `PublicKey.FromHex()` static method.  [[#2709], [#3044]]
  -  Added `PublicKey.ToHex()` method.  [[#2709], [#3044]]
  -  (Libplanet.Net) Added `Gossip.PublishMessage()` method.  [[#3054], [#3060]]
+ -  (@planetarium/account) Added `Account.getPublicKey()` async method.
+    [[#3061]]
+ -  (@planetarium/account) Added `RawPrivateKey.getPublicKey()` async method.
+    [[#3061]]
+ -  (@planetarium/account-aws-kms) Added `AwsKmsAccount.getPublicKey()` async
+    method.  [[#3061]]
 
 ### Behavioral changes
 
@@ -143,6 +161,7 @@ To be released.
 [#3044]: https://github.com/planetarium/libplanet/pull/3044
 [#3054]: https://github.com/planetarium/libplanet/issues/3054
 [#3060]: https://github.com/planetarium/libplanet/pull/3060
+[#3061]: https://github.com/planetarium/libplanet/pull/3061
 [#3067]: https://github.com/planetarium/libplanet/pull/3067